Move batchTransfer function into exchange contract to reuse transfer manager logic
